refactor: Query, Command Controller 분리
